Introducing a change that causes reasoning to give an error.
In this case the change made objective `characterize-atmosphere` and
objective `characterize-liquid-ocean` have the same base:hasIdentifier
value. Since base:hasIdentifier is a `key` property of aspect
base:IdentifiedThing, which is a superclass of concept
mission:Objective, the reasoner now infers that these two names are
aliases to the same objective individual.

In addition, the first objective is asserted to base:aggregates the
second objective. This now means that the objective individual
base:aggregates itself. Given that base:aggregates is an `asymmetric`
relation (meaning if x aggregates y, then y cannot aggregate x), it
follows that there is a logical inconsistency here.